Allow sharing from the main page
I've mostly been developing this on Chrome OS and Android, both of which
have the "share" dialog integrated with the OS. I can "install" the
`reading.cceckman.com` application; whenever I "share" I can target it.

On Windows, the Chrome "share" button triggers a Windows dialog, which
isn't aware of the "installed" application.

For all the desktop / laptop platforms, it's ~easy to copy-paste a link
(though it would be nice to have it integrated). This adds a `<form>` to
the main web app page that does the same thing as hitting the
`share_target`.