Capture install prompts, try to detect installation
This has an attempt to prompt for installation whenever the app isn't
installed. However, it's not reliable- I can't seem to get the
"related_applications" bits to report reflexive installation, i.e.
installation of the web app itself.

I'm going to roll this backwards to "just handle the install-prompt"- we
only need to encourage ("here is why you should install") if the
installation is needed, and we can just rely on the browser to
only fire the install-prompt event if it isn't installed already.